# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ServerBee API",
    "description": "ServerBee VPS monitoring probe API. All responses are wrapped in `{\"data\": <value>}`. Errors return `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/server-groups": {
      "get": {
        "tags": ["server-groups"],
        "operationId": "list_server_groups",
        "responses": {
          "200": {
            "description": "List all server groups",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerGroup"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": ["server-groups"],
        "operationId": "create_server_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Group created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerGroup"
                }
              }
            }
          },
          "409": {
            "description": "Group name already exists"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/server-groups/{id}": {
      "put": {
        "tags": ["server-groups"],
        "operationId": "update_server_group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerGroup"
                }
              }
            }
          },
          "404": {
            "description": "Group not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": ["server-groups"],
        "operationId": "delete_server_group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server group ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted"
          },
          "404": {
            "description": "Group not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/servers/{id}": {
      "get": {
        "tags": ["servers"],
        "operationId": "get_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": ["servers"],
        "operationId": "update_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": ["servers"],
        "operationId": "delete_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server deleted"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateGroupRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "ServerGroup": {
        "type": "object",
        "required": ["id", "name", "weight", "created_at"],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ServerResponse": {
        "type": "object",
        "description": "Server response DTO — excludes sensitive fields (token_hash, token_prefix).",
        "required": [
          "id",
          "name",
          "weight",
          "hidden",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "agent_version": {
            "type": ["string", "null"]
          },
          "billing_cycle": {
            "type": ["string", "null"]
          },
          "country_code": {
            "type": ["string", "null"]
          },
          "cpu_arch": {
            "type": ["string", "null"]
          },
          "cpu_cores": {
            "type": ["integer", "null"],
            "format": "int32"
          },
          "cpu_name": {
            "type": ["string", "null"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "currency": {
            "type": ["string", "null"]
          },
          "disk_total": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "expired_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "group_id": {
            "type": ["string", "null"]
          },
          "hidden": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "ipv4": {
            "type": ["string", "null"]
          },
          "ipv6": {
            "type": ["string", "null"]
          },
          "kernel_version": {
            "type": ["string", "null"]
          },
          "mem_total": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "os": {
            "type": ["string", "null"]
          },
          "price": {
            "type": ["number", "null"],
            "format": "double"
          },
          "public_remark": {
            "type": ["string", "null"]
          },
          "region": {
            "type": ["string", "null"]
          },
          "remark": {
            "type": ["string", "null"]
          },
          "swap_total": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "traffic_limit": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "traffic_limit_type": {
            "type": ["string", "null"]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "virtualization": {
            "type": ["string", "null"]
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": ["string", "null"]
          },
          "weight": {
            "type": ["integer", "null"],
            "format": "int32"
          }
        }
      },
...
      "UpdateServerInput": {
        "type": "object",
        "properties": {
          "billing_cycle": {
            "type": ["string", "null"]
          },
          "currency": {
            "type": ["string", "null"]
          },
          "expired_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "group_id": {
            "type": ["string", "null"]
          },
          "hidden": {
            "type": ["boolean", "null"]
          },
          "name": {
            "type": ["string", "null"]
          },
          "price": {
            "type": ["number", "null"],
            "format": "double"
          },
          "public_remark": {
            "type": ["string", "null"]
          },
          "remark": {
            "type": ["string", "null"]
          },
          "traffic_limit": {
            "type": ["integer", "null"],
            "format": "int64"
          },
          "traffic_limit_type": {
            "type": ["string", "null"]
          },
          "weight": {
            "type": ["integer", "null"],
            "format": "int32"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "session_cookie": []
    },
    {
      "api_key": []
    },
    {
      "bearer_token": []
    }
  ],
...
}